pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "forge-arena"
version = "0.1.0"
description = "Adaptive RL training environment for AI oversight capability."
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "httpx>=0.27.0",
    "rouge-score>=0.1.2",
    "scikit-learn>=1.4.0",
    "numpy>=1.26.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    "openenv-core>=0.2.0",
]

[project.scripts]
server = "server.app:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-benchmark>=4.0.0",
    "httpx>=0.27.0",
    "mypy>=1.10.0",
    "ruff>=0.4.0",
    "types-pyyaml>=6.0.12",
]
training = [
    "torch>=2.3.0",
    "transformers>=4.41.0",
    "trl>=0.9.0",
    "bitsandbytes>=0.43.0",
    "datasets>=2.19.0",
    "accelerate>=0.30.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv.run]
script = "uvicorn forge_arena.main:app --host 0.0.0.0 --port 8000"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "ANN", "S", "B"]
ignore = ["ANN101", "ANN102", "S101"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
